checkCardtraderStock/__init__.py
import logging
import azure.functions as func
import os
import datetime
import time
import requests
import pytz
from azure.data.tables import TableServiceClient, UpdateMode
from azure.core.exceptions import ResourceNotFoundError

# Constants for table names - replace with actual names or environment variables
USERS_TABLE_NAME = "users" # Table listing all user IDs
TIMESTAMPS_TABLE_NAME = "userCheckTimestamps" # Table tracking last check time per user
BLUEPRINTS_TABLE_NAME = "blueprintscardtrader" # Table with Cardtrader blueprint IDs
USER_TABLE_PREFIX = "user" # Prefix for user-specific tables

# Cardtrader API settings - use environment variables
CARDTRADER_API_KEY = os.environ.get("CARDTRADER_API_KEY")
# Example URL, replace with the actual marketplace endpoint if different
CARDTRADER_MARKETPLACE_URL = "https://api.cardtrader.com/v2/marketplace/products/{blueprint_id}"

# Configuration
RATE_LIMIT_SECONDS = 1.1 # Slightly more than 1 second to be safe
CHECK_INTERVAL_HOURS = 24 # Check each user at most once per day

def get_cardtrader_session():
    """Creates a requests session with Cardtrader auth headers."""
    if not CARDTRADER_API_KEY:
        raise ValueError("CARDTRADER_API_KEY environment variable not set.")
    
    session = requests.Session()
    session.headers.update({
        'Authorization': f'Bearer {CARDTRADER_API_KEY}',
        'Accept': 'application/json'
    })
    return session

def main(timer: func.TimerRequest) -> None:
    utc_timestamp = datetime.datetime.now(pytz.utc).isoformat()

    if timer.past_due:
        logging.info('The timer is past due!')

    logging.info(f'Python timer trigger function ran at {utc_timestamp}')

    # --- Implementation Placeholder ---
    # 1. Get connection string
    # 2. Initialize TableServiceClient
    # 3. Get clients for users, timestamps, blueprints tables
    # 4. Get list of user IDs from USERS_TABLE_NAME
    # 5. Query TIMESTAMPS_TABLE_NAME for these users
    # 6. Determine next user to check based on CHECK_INTERVAL_HOURS and oldest timestamp
    # 7. If no user needs checking, log and return
    # 8. Get user's table client (USER_TABLE_PREFIX + user_id)
    # 9. Query all cards from user's table
    # 10. Initialize Cardtrader session
    # 11. Loop through cards:
    #     a. Find blueprint ID from BLUEPRINTS_TABLE_NAME
    #     b. If found:
    #         i.   Wait RATE_LIMIT_SECONDS
    #         ii.  Call Cardtrader API (CARDTRADER_MARKETPLACE_URL)
    #         iii. Parse response for stock status
    #         iv.  Update card entity in user table (cardtrader_stock=True/False)
    #     c. Else (blueprint not found):
    #         i.   Log warning
    #         ii.  Update card entity (cardtrader_stock=False)
    # 12. Update timestamp for the checked user in TIMESTAMPS_TABLE_NAME
    # 13. Add comprehensive error handling

    logging.info("checkCardtraderStock function execution finished.")
    # Add actual implementation logic here in the next steps.

    pass # Placeholder for actual logic
