requirements.txt
azure-functions
azure-data-tables>=12.4.0
azure-core>=1.24.0
azure-identity>=1.12.0
requests>=2.31.0
pytz>=2023.3 
